Add forecast and weather details

+++ WeatherApp.Model/Coordinate.cs
@@ -2,7 +2,7 @@
 {
     public class Coordinate
     {
-        public decimal lon { get; set; }
-        public decimal lat { get; set; }
+        public float lon { get; set; }
+        public float lat { get; set; }
     }
 }
